adding to cart has been processed

+++ BuyssyApp/data/entity/Products.swift
@@ -12,10 +12,10 @@ class Products: Codable {
     var ad: String?
     var resim: String?
     var kategori: String?
-    var fiyat: Double?
+    var fiyat: Int?
     var marka: String?
     
-    init(id: Int, ad: String, resim: String, kategori: String, fiyat: Double, marka: String) {
+    init(id: Int, ad: String, resim: String, kategori: String, fiyat: Int, marka: String) {
         self.id = id
         self.ad = ad
         self.resim = resim
